requirements.txt
opencv-python
torch
torchvision
numpy
Pillow
matplotlib
streamlit
scikit-image